requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
asttokens==3.0.0
black==24.10.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
google-auth==2.37.0
griffe==1.5.1
groq==0.13.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
jiter==0.8.2
json_repair==0.32.0
jsonpath-python==1.0.6
jupyter_client==8.6.3
jupyter_core==5.7.2
logfire-api==2.9.0
matplotlib-inline==0.1.7
mistralai==1.2.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
openai==1.58.1
packaging==24.2
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic-ai==0.0.14
pydantic-ai-slim==0.0.14
pydantic_core==2.27.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==308
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
python-dotenv==1.0.1
langchain-groq
langchain-google-genai
langchain
langchain_community
langgraph 
langsmith
langchain_groq
IPython
langchain_huggingface
langchain_pinecone
pinecone
phidata
phi==0.6.7
yfinance
duckduckgo-search